move categoryToUrlParam and urlParamToCategory into pageTreeCategories.tsx
